separate util into one module
